+++ src/Console/MakeDtoDocsCommans.php
@@ -2,8 +2,10 @@
 
 namespace Bfg\Dto\Console;
 
+use Bfg\Dto\Dto;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
 use Symfony\Component\Finder\SplFileInfo;
 
 class MakeDtoDocsCommans extends Command
@@ -26,7 +28,7 @@ class MakeDtoDocsCommans extends Command
         $classes = collect($files)->map(function ($file) {
             return $this->getClassNameFromFile($file);
         })->filter()->values()->map(function ($class) {
-            $this->updateClassDocBlockWithReflection($class['class'], $class['doc']);
+            $this->updateClassDocBlockWithReflection($class['class'], $class['doc'], $class['old_doc']);
             return $class['class'];
         })->count();
 
@@ -49,11 +51,35 @@ class MakeDtoDocsCommans extends Command
             return null;
         }
         $ref = new \ReflectionClass($class);
-        $doc = ($ref)->getDocComment() ?: '';
+        $doc = ($ref)->getDocComment() ?: "";
         $extendProperty = $ref->getProperty('extends');
         $extendProperty->setAccessible(true);
         $extends = $extendProperty->getValue();
+        if (method_exists($class, 'getConstructorParameters')) {
+            $parameters = $class::getConstructorParameters();
+        } else {
+            $parameters = [];
+        }
+
         $variables = [];
+
+        /** @var \ReflectionParameter $parameter */
+        foreach ($parameters as $parameter) {
+            $name = $parameter->getName();
+            $type = $parameter->getType();
+            $types = [];
+            if ($type instanceof \ReflectionUnionType) {
+                foreach ($type->getTypes() as $unionType) {
+                    $types[] = $unionType->getName();
+                }
+            } else {
+                $types[] = $type->getName();
+            }
+            $variables['lazy' . ucfirst(Str::camel($name))] = implode('|', array_map(function ($val) {
+                return class_exists($val) || enum_exists($val) ? "\\" . $val : $val;
+            }, $types));
+        }
+
         foreach ($extends as $key => $val) {
             if (is_array($val)) {
                 $variables[$key] = implode('|', array_map(function ($val) {
@@ -62,6 +88,25 @@ class MakeDtoDocsCommans extends Command
             } else {
                 $variables[$key] = class_exists($val) || enum_exists($val) ? "\\" . $val : $val;
             }
+
+            $variables['lazy'. ucfirst(Str::camel($key))] = $variables[$key];
+        }
+
+        $computed = collect($ref->getMethods())->filter(function (\ReflectionMethod $method) {
+            return !in_array($method->getName(), [
+                    '__construct', '__destruct', '__get', '__set', '__isset', '__unset', '__sleep', '__wakeup',
+                    '__serialize', '__unserialize', '__clone', '__toString', '__invoke', '__set_state', '__clone',
+                    '__debugInfo', '__serialize', '__unserialize', '__sleep', '__wakeup'
+                ]) && ! method_exists(Dto::class, $method->getName())
+                && ! str_starts_with($method->getName(), 'fromArray')
+                && ! str_starts_with($method->getName(), 'toArray')
+                && ! str_starts_with($method->getName(), 'with')
+                && ! str_starts_with($method->getName(), 'lazy');
+        })->values();
+
+        foreach ($computed as $method) {
+            $variables[$method->getName()] = 'mixed';
+            $variables['lazy'. ucfirst(Str::camel($method->getName()))] = 'mixed';
         }
 
         if (! $variables) {
@@ -70,105 +115,57 @@ class MakeDtoDocsCommans extends Command
 
         return [
             'class' => $class,
+            'old_doc' => $doc,
             'doc' => $this->updateDocBlock($doc, $variables),
         ];
     }
 
-    protected function updateClassDocBlockWithReflection(string $className, string $docBlock): void
+    protected function updateClassDocBlockWithReflection(string $className, string $docBlock, string $oldDoc): void
     {
-        // Используем рефлексию, чтобы получить файл класса
         $reflection = new \ReflectionClass($className);
         $fileName = $reflection->getFileName();
-
-        if (!$fileName || !is_file($fileName)) {
-            throw new \RuntimeException("Файл класса {$className} не найден.");
-        }
-
-        // Загружаем содержимое файла
         $code = file_get_contents($fileName);
 
-        if ($code === false) {
-            throw new \RuntimeException("Не удалось загрузить содержимое файла {$fileName}.");
-        }
-
-        // Регулярное выражение для поиска докблока
-        $docBlockPattern = '/\/\*\*.*?\*\//s';
-
-        if (preg_match($docBlockPattern, $code, $matches, PREG_OFFSET_CAPTURE)) {
-            // Если докблок существует, заменяем его
-            $startPos = $matches[0][1];
-            $endPos = $startPos + strlen($matches[0][0]);
-            $code = substr_replace($code, $docBlock, $startPos, $endPos - $startPos);
+        if ($oldDoc) {
+            $code = str_replace($oldDoc, $docBlock, $code);
         } else {
-            // Если докблока нет, добавляем новый перед классом
-            if (preg_match('/\bclass\s+\w+/', $code, $matches, PREG_OFFSET_CAPTURE)) {
-                $classPos = $matches[0][1];
-                $code = substr_replace($code, $docBlock . "\n", $classPos, 0);
-            } else {
-                throw new \RuntimeException("Не удалось найти объявление класса {$className}.");
-            }
+            $code = preg_replace('/(class\s+' . $reflection->getShortName() . '.*?{)/s', "{$docBlock}\n$0", $code);
         }
 
-        // Сохраняем обновлённый код в файл
-        if (file_put_contents($fileName, $code) === false) {
-            throw new \RuntimeException("Не удалось записать изменения в файл {$fileName}.");
-        }
+        file_put_contents($fileName, $code);
     }
 
-    protected function updateDocBlock(string $code, array $properties): string
+    /**
+     * @param  string  $code
+     * @param  array  $properties
+     * @return string
+     */
+    protected function updateDocBlock(string $code, array $properties): string|array
     {
-        $docBlockPattern = '/\/\*\*.*?\*\//s';
-
-        $generateDocBlock = function (array $properties) use ($code): string {
-            if (empty($properties)) {
-                return "/**\n * @property mixed \$autoGeneratedProperty\n */";
-            }
-
-            $docBlock = "/**\n";
-            foreach ($properties as $property => $type) {
-                if (! str_contains($code, "\${$property}")) {
-                    $docBlock .= " * @property {$type} \${$property}\n";
-                }
-            }
-            $docBlock .= " */";
-
-            return $docBlock;
-        };
-
-        $newDocBlock = $generateDocBlock($properties);
-
-        if (preg_match($docBlockPattern, $code, $matches, PREG_OFFSET_CAPTURE)) {
-            $existingDocBlock = $matches[0][0];
-            $startPos = $matches[0][1];
-            $endPos = $startPos + strlen($existingDocBlock);
-
-            if (!empty($properties)) {
-                $updatedDocBlock = $existingDocBlock;
-                foreach ($properties as $property => $type) {
-                    if (! str_contains($code, "\${$property}")) {
-                        $line = " * @property {$type} \${$property}";
-                        if (!str_contains($updatedDocBlock, $line)) {
-                            $updatedDocBlock = substr_replace(
-                                $updatedDocBlock,
-                                "\n{$line}",
-                                strrpos($updatedDocBlock, "\n"),
-                                0
-                            );
-                        }
-                    }
-                }
-            } else {
-                $updatedDocBlock = $existingDocBlock;
-            }
-
-            return substr_replace($code, $updatedDocBlock, $startPos, $endPos - $startPos);
+        $lines = explode("\n", $code);
+        if (! count($lines)) {
+            $lines = ['/**', ' */'];
         } else {
-            if (preg_match('/\bclass\s+\w+/', $code, $matches, PREG_OFFSET_CAPTURE)) {
-                $classPos = $matches[0][1];
-                return substr_replace($code, $newDocBlock . "\n", $classPos, 0);
-            } else {
-                throw new \Exception("Cannot find class declaration in the file.");
+            if (! isset($lines[0]) || trim($lines[0]) !== '/**') {
+                $lines[0] = '/**';
+            }
+            $endIndex = count($lines) - 1;
+            $endIndex = $endIndex === 0 ? 1 : $endIndex;
+            if (! isset($lines[$endIndex]) || trim($lines[$endIndex]) !== ' */') {
+                $lines[$endIndex] = ' */';
             }
         }
+
+        $start = $lines[0];
+        $end = $lines[count($lines) - 1];
+        $lines = array_slice($lines, 1, count($lines) - 2);
+
+        foreach ($properties as $property => $type) {
+            if (! str_contains($code, "\${$property}")) {
+                $lines[] = " * @property {$type} \${$property}";
+            }
+        }
+
+        return $start . "\n" . implode("\n", $lines) . "\n" . $end;
     }
 }